damage handling now hits within a continuous range after mulitplications
A skill that does 9-11 damage with a multiplier of 3 would give a discrete set of damage outputs of 27, 30, or 33. This change aims to address that and give a continuous output of 27-33.

Disables randomization in early parts of the skill damage calculation, and then applies them at the end after multiplications.